Add support for legacy Lyft RLS
Note that I (Shahriar) tried to use tools/proto_format for generating
v1 things under generated_api_shadow but it seems it's not suited for
downgrading rather it's useful for migration from v2 to v3 and
probably to v4 for future. So I manually place the legacy files under
generated_api_shadow.

Signed-off-by: Gabriel <gsagula@gmail.com>
Signed-off-by: Luke Shumaker <lukeshu@datawire.io>
Signed-off-by: Shahriar Rostami <srdsecond@gmail.com>